Improve admin UI/UX with navigation links (#78)
* fix board editing ui

* Improves admin UI/UX with navigation links

Adds "Back to" links in admin views for better navigation.
Renames "Board" to "Imageboard" for clarity in the dashboard and edit view.
Adds a Display Name attribute to the UserRoleIds property in the UserEditViewModel.

The changes aim to enhance the user experience for administrators by providing easier navigation and clearer labels.