speed up thread queries (#86)
* Adds LastBumpAt property to Thread entity

Adds a `LastBumpAt` property to the `Thread` entity, which
represents the last time a thread was "bumped" (i.e., had a new
non-sage post or a post with attachments).

fix

* Adds page execution time to footer

Implements an action filter to measure page execution time
and displays it in the footer. This provides insights into
page performance and helps identify potential bottlenecks.

* Adds original post to threads

Adds the ability to associate a thread with its original post.
This allows for easier retrieval and display of the initial post
when viewing a thread.

* change db schema

* fix tests