move BoardName to config (#87)
The name property of the Board entity is no longer required.

This change addresses the removal of the board name as a core identifying attribute, simplifying the data model. The Board Name is now sourced from configuration.

Removes the edit board functionality since the board name is not an entity.